# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: QUANTEX API - AI-powered technician booking platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # Service Categories
...
  /technicians:
    get:
      operationId: listTechnicians
      tags: [technicians]
      summary: List technicians
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: available
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Technicians list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Technician"
    post:
      operationId: createTechnicianProfile
      tags: [technicians]
      summary: Create technician profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TechnicianInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technician"
  /technicians/{id}:
    get:
      operationId: getTechnician
      tags: [technicians]
      summary: Get technician by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Technician found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technician"
        "404":
          description: Not found
    patch:
      operationId: updateTechnician
      tags: [technicians]
      summary: Update technician profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TechnicianUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technician"
  /technicians/{id}/reviews:
    get:
      operationId: getTechnicianReviews
      tags: [technicians]
      summary: Get reviews for a technician
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reviews list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"

  # Bookings
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [customer, technician]
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]

    # Service Category
...
    Technician:
      type: object
      required: [id, userId, name, skills, rating, reviewCount, isAvailable, completedJobs, hourlyRate, responseTime]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        rating:
          type: number
        reviewCount:
          type: integer
        isAvailable:
          type: boolean
        completedJobs:
          type: integer
        hourlyRate:
          type: number
        responseTime:
          type: string
        distance:
          type: ["number", "null"]
        categoryIds:
          type: array
          items:
            type: integer
    TechnicianInput:
      type: object
      required: [skills, hourlyRate, responseTime]
      properties:
        bio:
          type: string
        skills:
          type: array
          items:
            type: string
        hourlyRate:
          type: number
        responseTime:
          type: string
        categoryIds:
          type: array
          items:
            type: integer
    TechnicianUpdate:
      type: object
      properties:
        bio:
          type: string
        skills:
          type: array
          items:
            type: string
        hourlyRate:
          type: number
        responseTime:
          type: string
        isAvailable:
          type: boolean
        categoryIds:
          type: array
          items:
            type: integer

    # Booking
...
    Review:
      type: object
      required: [id, customerId, technicianId, bookingId, rating, comment, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        technicianId:
          type: integer
        bookingId:
          type: integer
        rating:
          type: integer
        comment:
          type: string
        customerName:
          type: ["string", "null"]
        technicianName:
          type: ["string", "null"]
        createdAt:
          type: string
...