# lib/api-zod/src/generated/types/technicianUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * QUANTEX API - AI-powered technician booking platform
 * OpenAPI spec version: 0.1.0
 */

export interface TechnicianUpdate {
  bio?: string;
  skills?: string[];
  hourlyRate?: number;
  responseTime?: string;
  isAvailable?: boolean;
  categoryIds?: number[];
}